// src/lib/api/generated/model/fileServiceUpdateFileBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */

export interface FileServiceUpdateFileBody {
  content: string;
  message?: string;
  sha?: string;
  branch?: string;
  committerName?: string;
  committerEmail?: string;
}