// src/lib/api/generated/model/v1CreateFileResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Aisphere Hub API
 * OpenAPI spec version: v1
 */

export interface V1CreateFileResponse {
  commitMessage?: string;
  commitSha?: string;
  content?: string;
  encoding?: string;
  lastModified?: string;
  name?: string;
  path?: string;
  ref?: string;
  sha?: string;
  size?: string;
}