// src/lib/api/generated/model/v1FileInfo.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */

export interface V1FileInfo {
  name?: string;
  path?: string;
  type?: string;
  size?: string;
  mode?: string;
  sha?: string;
  lastModified?: string;
}